No plus sign in path
When rebar3 encounters multiple profile strings, such as "ci" and "test",
it concatenates them with a "+", yielding "ci+test". I've encountered issues
where CircleCI and IntelliJ attempt to serve the raw path and fail,
interpeting the "+" as a URL-encoded space. For simplicity, this PR
uses "_" instead of "+" to concatenate mutliple profile strings.