Add experimental support for ct --retry option
This commit adds a common test hook along with the cth_readable stuff
whose role is to track failing test cases, and create a test
specification out of them.

The test specification is dumped on disk at
_build/<profile>/logs/retry.spec and can be accessed by calling
'rebar3 ct --retry'. This will auto-load the spec file if it can be
found and re-run the failing cases.

If any other argument is found on the list specifying tests, the
'--retry' argument is ignored.

All code for this is marked as experimental in case we end up (keeping
and then) dropping the feature.