Restore default templates in bootstrapped rebar3
When allowing Dialyzer to work internally, we moved a bunch of
config (such as no_debug_info) to the prod profile, but the escript
included files remained locked onto the default profile.

Because the bootstrapping phase now happened in a prod run, the priv/
dirs were never created for the default profile unless a prior run
existed, which hid the bug from us.

This patch overrides the path on the prod profile so that we don't rely
on accidental leftovers for things to work on a bootstrap phase as a
dev.