Reload logger config in shell
This requires some fancy dynamic work since the logger is started as
part of the kernel and we lost the sys.config from users when working
from there.

We start conservatively by making it an optional thing, turning it on
only where we know it to be safe.

The changes are applied _after_ having loaded the rest of configs so if
an inoffensive error happens, the shell works (with a bad error message)
rather than plain exploding.