reset default profile app plugins to empty if at the top level (#1885)
top level default profile plugins are installed in the run_aux
function. This commit removes the plugins from the plugin opts
of an application that is also at the top level and skips the
default profile when installing top level plugins in
project_apps_install so they aren't handled twice.

Additionally, before handling a plugin the state's list of
known plugin apps is checked and the plugin is skipped if it
has already been handled.