(#1833) Include one more level in verify_config_format
top level tuples in the 'rebar.config' file are
already verified to be a tuple of size two.
If something else is specified, rebar3 will exit
with an error message. But in the next level,
users can enter empty tuples, and tuples of
size one and rebar3 will crash later on.

This commit fixes strange 'rebar.config'
definitions in the next level. Perhaps it should
verify even deeper levels?