Fix path coverage loading artifacts
The old path handling routine would force loading a bunch of
applications, but it did not necessarily unload them aside from specific
calls that the other test handlers used to need to do, but no longer
needed under the updated path management of 3.7.0.

The leftover calls in the cover compilation bits caused some apps to be
loaded and to remain loaded before tests started.