Attempt at a demo diagnostic mode.
Diagnostic mode is intended to separate the two use cases of DEBUG=1
mode of rebar3:

1. The user trying to understand what is going on
2. A maintainer trying to dig into what leads to a given behaviour

The scope for both should be different, if only because the user's
interactions with the system should only be done through:

- rebar.config
- the files on disk
- the CLI commands
- a superficial understanding of how the tool works

The first three points tell us that we should provide user debug
output first and foremost on these topics, since those are the
interaction points users will have with the tool.

Anything that _requires_ more than superficial understanding of how the
tool works in the output is meaningless, or actively harming
understanding. This kind of output should only point towards gently
nudging the user towards understanding how the tool operates without
digging in the code (i.e. outputting the implicit commands yours depends
on)

For a maintainer, this is different. What we want is raw information
that we can rattach and reinterpret in the context of more specific
knowledge of the internals, with support from the source code if
necessary.

This is a proof of concept that keeps the former in DEBUG=1, and places
the advanced stuff in DIAGNOSTIC=1, to invent a term.