Disable the default logger handler in shell if required
To avoid a conflict the Kernel's default handler have to be disabled
before the new default handler is added

    {kernel,
     [{logger,
       [{handler, default, undefined}
       ]}
     ]},
    {my_app,
     [{logger,
       [{handler, default, my_handler, #{}}
       ]}
     ]}

To support this behavior in shell it's necessary to process the
handler-default-undefined tuple.

It's not clear what to do with the simple handler (logger_simple_h)
however.  It's added at the start and removed if a default handler was
added.  Should it be added in reread_logger_config/1 if there's no
default handler?