Change cwd for each erl file compiled and back
Since the Erlang compiler relies on the current working directory to
find the .hrl files in -include or -include_lib instructions, and that
rebar3 sticks to the project root, then include files whose name clash
across a dep and a root project can end up crossing each-other.

The performance cost of this might be difficult to accept.