remove warning when booting applications in the shell
This warning is emitted every single time one starts a shell with applications
configured for automatic startup (e.g. `{shell, [{apps, [cowboy]}]}`).

Infortunately it does not make any sense: people who use this feature need it
to interact with their application or library during development, and are
therefore subjected to the warning every single time they start the shell even
though this is the right way to use it.

While the shell is not to be used to run applications in productions, those
who do it won't be stopped by a warning.

Ultimately, the warning annoys people who use the feature as intended, and do
nothing to stop people who abuse it. Thus I believe it should be removed.

See https://github.com/erlang/rebar3/issues/2206.

close #2206