Fix bootstrap crashes when compilers get modified
When recompiling rebar3 itself, the path swaps and cleaning removes the
modules loaded from _build/bootstrap, but the VM still manages to
discover those in _build/prod from previous builds of the escript when
building the new escript, which causes weird failures when compilers get
modified between releases.

The fix is to just clear _build/prod as part of the init of the
bootstrap script, the same way we already clear _build/bootstrap.